Support symlinks for activation routine.
When symlinking the BPGH plugin, the plugin's installation routine on
activation does not run.

The problem is due to WP's register_activation_hook() function, which does
not support symlinks.

This commit uses WP's 'activate_{PLUGIN_BASENAME}' hook, which works with
symlinks since we have to calculate the plugin basename ourselves.